+++ openstates/va/bills.py
@@ -1,89 +1,104 @@
 import re
+import pytz
 import datetime
-from collections import defaultdict
+import collections
 
-from billy.scrape.bills import BillScraper, Bill
-from billy.scrape.votes import Vote
-import lxml.html
+from spatula import Page, Spatula
+from pupa.scrape import Scraper, Bill, VoteEvent
 
+tz = pytz.timezone('America/New_York')
 
 BASE_URL = 'http://lis.virginia.gov'
+URL_PATTERNS = {
+    'list': '/cgi-bin/legp604.exe?{}+lst+ALL',
+    'summary': '/cgi-bin/legp604.exe?{}+sum+{}',
+    'sponsors': '/cgi-bin/legp604.exe?{}+mbr+{}',
+    'subjects': '/cgi-bin/legp604.exe?{}+sbj+SBJ',
+}
+ACTION_CLASSIFIERS = (
+    ('Approved by Governor', 'executive-signature'),
+    ('\s*Amendment(s)? .+ agreed', 'amendment-passage'),
+    ('\s*Amendment(s)? .+ withdrawn', 'amendment-withdrawal'),
+    ('\s*Amendment(s)? .+ rejected', 'amendment-failure'),
+    ('Subject matter referred', 'referral-committee'),
+    ('Rereferred to', 'referral-committee'),
+    ('Referred to', 'referral-committee'),
+    ('Assigned ', 'referral-committee'),
+    ('Reported from', 'committee-passage'),
+    ('Read third time and passed', ['passage', 'reading-3']),
+    ('Read third time and agreed', ['passage', 'reading-3']),
+    ('Passed (Senate|House)', 'passage'),
+    ('Read third time and defeated', 'failure'),
+    ('Presented', 'introduction'),
+    ('Prefiled and ordered printed', 'introduction'),
+    ('Read first time', 'reading-1'),
+    ('Read second time', 'reading-2'),
+    ('Read third time', 'reading-3'),
+    ('Senators: ', None),
+    ('Delegates: ', None),
+    ('Committee substitute printed', None),
+    ('Bill text as passed', None),
+    ('Acts of Assembly', None),
+)
 
+class SubjectPage(Page, Spatula):
+    def handle_page(self):
+        subjects = collections.defaultdict(list)
+        for link in self.doc.xpath('//ul[@class="linkSect"]/li/a'):
+            for bill_id in self.scrape_page(SubjectBillListPage, url=link.get('href')):
+                subjects[bill_id].append(link.text)
+        return subjects
 
-class VABillScraper(BillScraper):
-    jurisdiction = 'va'
+class SubjectBillListPage(Page, Spatula):
+    def handle_page(self):
+        for bill in self.doc.xpath('//ul[@class="linkSect"]/li'):
+            link = bill.getchildren()[0]
+            yield str(link.text_content())
+        next_url = self.doc.xpath('//a/b[text()="More..."]/../@href')
+        if next_url:
+            yield from self.scrape_page_items(SubjectBillListPage, url=next_url[0])
 
-    # There's a weird catch-all for numerals after the dash in the Yes
-    # count. That's because we've actually encountered this.
-    vote_strip_re = re.compile(r'(.+)\((\d+)-[\d]*Y (\d+)-N(?: (\d+)-A)?\)')
-    actor_map = {'House': 'lower', 'Senate': 'upper', 'Governor': 'governor',
-                 'Conference': 'conference'}
+class BillListPage(Page, Spatula):
+    def handle_page(self):
+        bills = self.doc.xpath('//ul[@class="linkSect"]/li')
+        for bill in bills:
+            link = bill.getchildren()[0]
+            bill_id = str(link.text_content())
 
-    _action_classifiers = (
-        ('Approved by Governor', 'governor:signed'),
-        ('\s*Amendment(s)? .+ agreed', 'amendment:passed'),
-        ('\s*Amendment(s)? .+ withdrawn', 'amendment:withdrawn'),
-        ('\s*Amendment(s)? .+ rejected', 'amendment:failed'),
-        ('Subject matter referred', 'committee:referred'),
-        ('Rereferred to', 'committee:referred'),
-        ('Referred to', 'committee:referred'),
-        ('Assigned ', 'committee:referred'),
-        ('Reported from', 'committee:passed'),
-        ('Read third time and passed', ['bill:passed', 'bill:reading:3']),
-        ('Read third time and agreed', ['bill:passed', 'bill:reading:3']),
-        ('Passed (Senate|House)', 'bill:passed'),
-        ('Read third time and defeated', 'bill:failed'),
-        ('Presented', 'bill:introduced'),
-        ('Prefiled and ordered printed', 'bill:introduced'),
-        ('Read first time', 'bill:reading:1'),
-        ('Read second time', 'bill:reading:2'),
-        ('Read third time', 'bill:reading:3'),
-        ('Senators: ', None),
-        ('Delegates: ', None),
-        ('Committee substitute printed', None),
-        ('Bill text as passed', None),
-        ('Acts of Assembly', None),
-    )
+            if not bill_id.startswith('S') or bill_id.startswith('H'):
+                continue
 
-    link_xpath = '//ul[@class="linkSect"]/li/a'
+            # create a bill
+            desc = bill.xpath('text()')[0].strip()
+            chamber = {
+                'H': 'lower',
+                'S': 'upper',
+            }[bill_id[0]]
+            bill_type = {'B': 'bill',
+                            'J': 'joint resolution',
+                            'R': 'resolution'}[bill_id[1]]
+            bill = Bill(bill_id, self.kwargs['session'], desc,
+                        chamber=chamber, classification=bill_type)
 
-    def _accept_response(self, response):
-        # check for rate limit pages
-        normal = super(VABillScraper, self)._accept_response(response)
-        return (normal and
-            'Sorry, your query could not be processed' not in response.text
-            and 'the source database is temporarily unavailable' not in response.text)
+            bill_url = link.get('href')
+            sponsor_url = BASE_URL + URL_PATTERNS['sponsors'].format(
+                self.kwargs['session'],
+                bill_id.replace(' ', ''),
+            )
 
-    def _get_page_bills(self, issue_name, href):
-        issue_html = self.get('http://lis.virginia.gov' + href,
-                                  retry_on_404=True).text
-        idoc = lxml.html.fromstring(issue_html)
-        for ilink in idoc.xpath(self.link_xpath):
-            self.subject_map[ilink.text].append(issue_name)
+            list(self.scrape_page_items(BillSponsorPage, url=sponsor_url, obj=bill))
+            list(self.scrape_page_items(BillDetailPage, url=bill_url, obj=bill))
+            bill.subject = self.kwargs['subjects'][bill_id]
+            bill.add_source(bill_url)
+            yield bill
 
-        more_links = idoc.xpath('//a/b[text()="More..."]/../@href')
-        if more_links:
-            self._get_page_bills(issue_name, more_links[0])
+        next_url = self.doc.xpath('//a/b[text()="More..."]/../@href')
+        if next_url:
+            yield from self.scrape_page_items(BillListPage, url=next_url[0], **self.kwargs)
 
-    def _build_subject_map(self):
-        url = 'http://lis.virginia.gov/cgi-bin/legp604.exe?%s+sbj+SBJ' % self.site_id
-        self.subject_map = defaultdict(list)
-
-        # loop over list of all issue pages
-        html = self.get(url).text
-        doc = lxml.html.fromstring(html)
-        for link in doc.xpath(self.link_xpath):
-            # get bills from page
-            self._get_page_bills(link.text, link.get('href'))
-
-    def _fetch_sponsors(self, bill):
-        url = "http://lis.virginia.gov/cgi-bin/legp604.exe?%s+mbr+%s" % (
-            self.site_id, bill['bill_id'].replace(' ', ''))
-
-        html = self.get(url, retry_on_404=True).text
-        doc = lxml.html.fromstring(html)
-
-        for slist in doc.xpath('//ul[@class="linkSect"]'):
+class BillSponsorPage(Page, Spatula):
+    def handle_page(self):
+        for slist in self.doc.xpath('//ul[@class="linkSect"]'):
             # note that first ul is origin chamber
             for sponsor in slist.xpath('li'):
                 name = sponsor.text_content().strip()
@@ -95,80 +110,46 @@ class VABillScraper(BillScraper):
                     type = 'cosponsor'
                 else:
                     type = 'cosponsor'
-                bill.add_sponsor(type, name)
+                self.obj.add_sponsorship(name, type, 'person', type == 'primary')
+                yield self.obj
 
-    def _split_vote(self, block):
-        if block:
-            block = block[0].text.replace('\r\n', ' ')
+class BillDetailPage(Page, Spatula):
 
-            pieces = block.split('--')
-            # if there are only two pieces, there are no abstentions
-            if len(pieces) <= 2:
-                return []
-            else:
-                # lookahead and don't split if comma precedes initials
-                # Also, Bell appears as Bell, Richard B. and Bell, Robert P.
-                # and so needs the lookbehind assertion.
-                return [x.strip() for x in re.split('(?<!Bell), (?!\w\.\w?\.?)', pieces[1]) if x.strip()]
-        else:
-            return []
+    vote_strip_re = re.compile(r'(.+)\((\d+)-[\d]*Y (\d+)-N(?: (\d+)-A)?\)')
+    actor_map = {'House': 'lower', 'Senate': 'upper', 'Governor': 'governor',
+                 'Conference': 'conference'}
 
-    def _parse_vote(self, vote, url):
-        url = BASE_URL + url
-
-        html = self.get(url, retry_on_404=True).text
-        doc = lxml.html.fromstring(html)
-
-        yeas = doc.xpath('//p[contains(text(), "YEAS--")]')
-        nays = doc.xpath('//p[contains(text(), "NAYS--")]')
-        # We capture "other" types of votes separately just in case we
-        # want to have the granularity later.
-        rule36 = doc.xpath('//p[contains(text(), "RULE 36--")]')
-        abstaining = doc.xpath('//p[contains(text(), "ABSTENTIONS--")]')
-        notvoting = doc.xpath('//p[contains(text(), "NOT VOTING--")]')
-
-        map(vote.yes, self._split_vote(yeas))
-        map(vote.no, self._split_vote(nays))
-        # Flattening all types of other votes into a single list.
-        other_votes = []
-        map(other_votes.extend, (self._split_vote(rule36), self._split_vote(abstaining),
-            self._split_vote(notvoting)))
-        map(vote.other, other_votes)
-
-    def _scrape_bill_details(self, url, bill):
-        html = self.get(url, retry_on_404=True).text
-        doc = lxml.html.fromstring(html)
-
-        # summary sections
-        summary = doc.xpath('//h4[starts-with(text(), "SUMMARY")]/following-sibling::p/text()')
-        if summary and summary[0].strip():
-            bill['summary'] = summary[0].strip()
+    def handle_page(self):
+        # TODO(jmcarp) Restore summary
+        # summary = self.doc.xpath('//h4[starts-with(text(), "SUMMARY")]/following-sibling::p/text()')
+        # if summary and summary[0].strip():
+        #     self.obj['summary'] = summary[0].strip()
 
         # versions
-        for va in doc.xpath('//h4[text()="FULL TEXT"]/following-sibling::ul[1]/li/a[1]'):
+        for va in self.doc.xpath('//h4[text()="FULL TEXT"]/following-sibling::ul[1]/li/a[1]'):
 
             # 11/16/09 \xa0House: Prefiled and ordered printed; offered 01/13/10 10100110D
             date, desc = va.text.split(u' \xa0')
             desc.rsplit(' ', 1)[0]              # chop off last part
             link = va.get('href')
             if 'http' not in link:
-                link = '{}{}'.format(BASE_URL,link)
+                link = '{}{}'.format(BASE_URL, link)
             date = datetime.datetime.strptime(date, '%m/%d/%y')
+            date = tz.localize(date)
 
             # budget bills in VA are searchable but no full text available
             if '+men+' in link:
                 self.warning('not adding budget version, bill text not available')
             else:
                 # VA duplicates reprinted bills, lets keep the original name
-                bill.add_version(desc, link, date=date,
-                                 mimetype='text/html',
-                                 on_duplicate='use_old')
+                self.obj.add_version_link(desc, link, date=date,
+                                          media_type='text/html',
+                                          on_duplicate='ignore')
 
         # actions
         cached_vote = None
         cached_action = None
-        for ali in doc.xpath('//h4[text()="HISTORY"]/following-sibling::ul[1]/'
-            'li'):
+        for ali in self.doc.xpath('//h4[text()="HISTORY"]/following-sibling::ul[1]/li'):
             vote = None
 
             date, action = ali.text_content().split(u' \xa0')
@@ -181,6 +162,7 @@ class VABillScraper(BillScraper):
 
             actor = self.actor_map[actor]
             date = datetime.datetime.strptime(date.strip(), '%m/%d/%y')
+            date = tz.localize(date)
 
             # if action ends in (##-Y ##-N) remove that part
             vrematch = self.vote_strip_re.match(action)
@@ -215,107 +197,133 @@ class VABillScraper(BillScraper):
                 # identical vote counts in the next line.
                 if cached_vote is None:
                     cached_action = action
-                    cached_vote = Vote(actor, date, vote_action, y > n, y, n,
-                        o)
+                    cached_vote = VoteEvent(
+                        start_date=date,
+                        chamber=actor,
+                        motion_text=vote_action,
+                        result='pass' if y > n else 'fail',
+                        classification='passage',
+                        bill=self.obj,
+                    )
+                    cached_vote.set_count('yes', y)
+                    cached_vote.set_count('no', n)
+                    cached_vote.set_count('other', o)
                     if vote_url:
-                        cached_vote.add_source(BASE_URL + vote_url[0])
+                        cached_vote.add_source(vote_url[0])
                     continue
                 elif cached_vote is not None:
                     if vote_action.startswith(u'VOTE:'):
+                        counts = {count['option']: count['value'] for count in cached_vote.counts}
                         if (vote_url
-                            and cached_vote['yes_count'] == y
-                            and cached_vote['no_count'] == n
-                            and cached_vote['other_count'] == o):
+                            and counts['yes'] == y
+                            and counts['no'] == n
+                            and counts['other'] == o):
                             vote = cached_vote
-                            self._parse_vote(vote, vote_url[0])
-                            vote.add_source(BASE_URL + vote_url[0])
+                            list(self.scrape_page_items(VotePage, url=vote_url[0], obj=vote))
+                            vote.add_source(vote_url[0])
                             action = cached_action
-                    elif cached_vote['motion'].startswith('VOTE:'):
-                        if (cached_vote['yes_count'] == y
-                            and cached_vote['no_count'] == n
-                            and cached_vote['other_count'] == o):
+                    elif cached_vote.motion_text.startswith('VOTE:'):
+                        counts = {count['option']: count['value'] for count in cached_vote.counts}
+                        if (counts['yes'] == y
+                            and counts['no'] == n
+                            and counts['other'] == o):
                             vote = cached_vote
                             vote['motion'] = vote_action
                     else:
                         # Cached vote doesn't match up to the current
                         # one. Save, then cache the current vote to
                         # begin the next search.
-                        bill.add_vote(cached_vote)
-                        cached_vote = Vote(actor, date, vote_action, y > n, y,
-                            n, o)
+                        yield cached_vote
+                        cached_vote = VoteEvent(
+                            start_date=date,
+                            chamber=actor,
+                            motion_text=vote_action,
+                            result='pass' if y > n else 'fail',
+                            classification='passage',
+                            bill=self.obj,
+                        )
+                        cached_vote.set_count('yes', y)
+                        cached_vote.set_count('no', n)
+                        cached_vote.set_count('other', o)
                         if vote_url:
-                            cached_vote.add_source(BASE_URL + vote_url[0])
+                            cached_vote.add_source(vote_url[0])
                         cached_action = action
                         continue
 
                 if vote is None:
                     raise ValueError('Cannot save an empty vote.')
-                #vote.validate()
-                bill.add_vote(vote)
+                yield vote
             else:
                 # If this action isn't a vote, but the last one was,
                 # there's obviously no additional vote data to match.
                 # Go ahead and save the cached data.
                 if cached_vote is not None:
-                    bill.add_vote(cached_vote)
+                    yield cached_vote
 
             cached_vote = cached_action = None
 
             # categorize actions
-            for pattern, atype in self._action_classifiers:
+            for pattern, atype in ACTION_CLASSIFIERS:
                 if re.match(pattern, action):
                     break
             else:
-                atype = 'other'
+                atype = None
 
             # if matched a 'None' atype, don't add the action
+            chamber = None  # TODO: populate
             if atype:
-                bill.add_action(actor, action, date, type=atype)
+                self.obj.add_action(action, date, organization=actor, chamber=chamber, classification=atype)
 
+class VotePage(Page):
+    def handle_page(self):
+        yeas = self.doc.xpath('//p[contains(text(), "YEAS--")]')
+        nays = self.doc.xpath('//p[contains(text(), "NAYS--")]')
+        # We capture "other" types of votes separately just in case we
+        # want to have the granularity later.
+        rule36 = self.doc.xpath('//p[contains(text(), "RULE 36--")]')
+        abstaining = self.doc.xpath('//p[contains(text(), "ABSTENTIONS--")]')
+        notvoting = self.doc.xpath('//p[contains(text(), "NOT VOTING--")]')
 
-    def scrape(self, chamber, session):
-        self.user_agent = 'openstates +mozilla'
-        # internal id for the session, store on self so all methods have access
-        self.site_id = self.metadata['session_details'][session]['site_id']
+        for name in self._split_vote(yeas):
+            self.obj.vote('yes', name)
+        for name in self._split_vote(nays):
+            self.obj.vote('no', name)
+        # Flattening all types of other votes into a single list.
+        other_votes = []
+        map(other_votes.extend, (self._split_vote(rule36), self._split_vote(abstaining),
+            self._split_vote(notvoting)))
+        for name in other_votes:
+            self.obj.vote('other', name)
+        yield self.obj
 
-        self._build_subject_map()
+    def _split_vote(self, block):
+        if block:
+            block = block[0].text.replace('\r\n', ' ')
 
-        # used for skipping bills from opposite chamber
-        start_letter = 'H' if chamber == 'lower' else 'S'
+            pieces = block.split('--')
+            # if there are only two pieces, there are no abstentions
+            if len(pieces) <= 2:
+                return []
+            else:
+                # lookahead and don't split if comma precedes initials
+                # Also, Bell appears as Bell, Richard B. and Bell, Robert P.
+                # and so needs the lookbehind assertion.
+                return [x.strip() for x in re.split('(?<!Bell), (?!\w\.\w?\.?)', pieces[1]) if x.strip()]
+        else:
+            return []
 
-        url = 'http://lis.virginia.gov/cgi-bin/legp604.exe?%s+lst+ALL' % self.site_id
+class VaBillScraper(Scraper, Spatula):
+    def scrape(self, session=None):
+        if not session:
+            session = self.jurisdiction.legislative_sessions[-1]['identifier']
+            self.info('no session specified, using', session)
+        url = BASE_URL + URL_PATTERNS['list'].format(session)
+        subjects = self.scrape_page(SubjectPage, url=BASE_URL + URL_PATTERNS['subjects'].format(session))
+        yield from self.scrape_page_items(BillListPage, url=url, session=session, subjects=subjects)
 
-        while url:
-            html = self.get(url, retry_on_404=True).text
-            doc = lxml.html.fromstring(html)
-
-            url = None  # no more unless we encounter 'More...'
-
-            bills = doc.xpath('//ul[@class="linkSect"]/li')
-            for bill in bills:
-                link = bill.getchildren()[0]
-                bill_id = str(link.text_content())
-
-                # check if this is the 'More...' link
-                if bill_id.startswith('More'):
-                    url = BASE_URL + link.get('href')
-
-                # skip bills from the other chamber
-                elif not bill_id.startswith(start_letter):
-                    continue
-
-                else:
-                    # create a bill
-                    desc = bill.xpath('text()')[0].strip()
-                    bill_type = {'B': 'bill',
-                                 'J': 'joint resolution',
-                                 'R': 'resolution'}[bill_id[1]]
-                    bill = Bill(session, chamber, bill_id, desc,
-                                type=bill_type)
-
-                    bill_url = BASE_URL + link.get('href')
-                    self._fetch_sponsors(bill)
-                    self._scrape_bill_details(bill_url, bill)
-                    bill['subjects'] = self.subject_map[bill_id]
-                    bill.add_source(bill_url)
-                    self.save_bill(bill)
+    def accept_response(self, response):
+        # check for rate limit pages
+        normal = super().accept_response(response)
+        return (normal and
+            'Sorry, your query could not be processed' not in response.text
+            and 'the source database is temporarily unavailable' not in response.text)
